+++ docs/changes.rst
@@ -11,6 +11,7 @@ To be released.
 - Added :option:`--trusted-proxy <geofront-server --trusted-proxy>` option
   to :program:`goefront-server` command.  It's useful when the server is
   run behind a reverse proxy.
+- Added token no-op API: :http:get:`/tokens/(token_id:token_id)/`.
 
 
 Version 0.1.0